Bug 1378239 - Change mozinfo 'coverage' flag to 'ccov' to avoid ambiguity. r=ahal
This patch renames the mozinfo flag 'coverage' to 'ccov' to avoid ambiguity in whether a test is being skipped for linux64-ccov or for linux64-jsdcov. It also removes the 'runtests.py' mozinfo hack and renames all occurrences of 'coverage' that are used for skipping tests in linux64-ccov.

MozReview-Commit-ID: IF2640bDQP7

UltraBlame original commit: 789ccac5e77dea00f2ff0f517f4795354164639b