Bug 1375231 - Make OptionValue.__eq__ more type aware; r=glandium
OptionValue and its derived classes are exposed to moz.configure
files. As the previous bug fix showed, it is really easy to
accidentally assume the type is a simple string value and do a
string compare against it.

To prevent this class of bugs, this commit adds additional type
awareness to OptionValue.__eq__.

We first check that the argument is a tuple (including any OptionValue
types). If not, we raise a TypeError because the comparison is
invalid. This is arguably a violation of __eq__. But since OptionValue
is exposed to the moz.configure sandbox and typing '==' will invoke
__eq__, we have to do something to prevent this foot gun.

The change also changes the comparison logic.

If we compare against a non-derived tuple instance, we do a tuple
compare. Otherwise, we fall back to the previous logic of
requiring an identical type then doing a tuple compare.

MozReview-Commit-ID: 7IVSL2Asg9j

UltraBlame original commit: 51a92a22d6d114db23a44ede30024177df322246