Bug 1378427 - move context menus around, fix flickering in dynamic portion of the overflow menu for photon, r=jaws
Apparently the hover state of the combined buttons does not interact well with the menus unless they
share the same containing popup/panel. We broke that condition in bug 1354078. The expedient thing
is to simply move the popup back, and move it back and forth if/when the photon pref is flipped.
When removing the pref, we can simplify this by always putting the menu in the overflow panel.

I also noticed that we use the toolbar context menu in the dynamic portion of the overflow panel.
This has the same problem, and to fix it I switched us to using the same (panel) context menu in
the photon case. Doing this in the non-photon case won't help because the context menu will be in
a separate panel (namely the old hamburger panel) entirely.

MozReview-Commit-ID: 4neHMukTzHA

UltraBlame original commit: 13302a007a78bc3060bffbdf7a29a7c14f25fa46