Bug 1379505: Less fishyness when resolving the style of the document element. r=heycam
Previous to these patches, the style resolution happening on [1] made the style
data stick on the element, so we'd never think it was the initial style, even if
it was.

This was wallpapering the fact that, if that was the initial style, we'd never
have another chance of traversing the document when [2] kicked in.

This somehow just happened to work, but is a very fishy way to get it to work.
Instead, call StyleDocument() properly _before_, and rely on the fact that it
will stop when it has a non-null binding, and only if it fails explicitly style
the children.

This fixes the few XBL test-cases with this patch series that exercise
-moz-bindings on the root element without the root being styled, and makes the
-moz-binding code more consistent in both places of the frame constructor.

[1]: http://searchfox.org/mozilla-central/rev/5dadcbe55b4ddd1e448c06c77390ff6483aa009b/layout/base/nsCSSFrameConstructor.cpp#2526
[2]: https://github.com/servo/servo/blob/3330653dc80cc8947af17f1989fb7fbf390c4d2f/components/style/traversal.rs#L439

MozReview-Commit-ID: HbjsD6nYsvX

UltraBlame original commit: 183625671488778f2321cb569ae57068be379010