Bug 1372069 - Part 1: Disable Geolocation when 'privacy.resistFingerprinting' is true. r=arthuredelstein,smaug
This patch disables Geolocation API when fingerprinting resistance is enabled.
The way we disable it is the same as how we disable this API for non-secure
origins that we will reject the request from this API and still keep this API
around.

MozReview-Commit-ID: 5D7Bf6Rplm8

UltraBlame original commit: a450b50c2222664f6f2e6bfff09b81274bc242a2