Bug 1380188 - Update GMP download fallback to Widevine CDM 970. r=mconley
We are pushing out a new Widevine CDM version, so we need to update the CDM
downloader to download the new version if the GMP AUS server is down.

Note that the existing code won't download a CDM/GMP using the fallback path if
a CDM/GMP is not already installed.


MozReview-Commit-ID: 6eF4QupkU77

UltraBlame original commit: 8ffc131e81692d70c479a0c72b68699adcb97b99