Bug 1381858 - Ensure we have a valid window handle before trying to wait, r=ato
If we switched to a window that is now closed (possible in a reftest
run, for example), we might end up calling execute_async_script
without a valid current window. The runner window is always supposed
to be open (and if it isn't then waiting won't make much sense
anyway), so switch to that before executing script in wait().

MozReview-Commit-ID: ERoN4sz5SuH

UltraBlame original commit: c3df1a01c190a85543f1fef04878ecef967857c5