Bug 1347963 - part 0 - use global qualification for GuardObjects.h macros; r=erahm
These macros can be used in cases where the `mozilla` namespace might
not refer to the toplevel `mozilla` namespace that was intended.  To
ensure that the macros always refer to the `mozilla` namespace in the
global namespace, use the appropriate qualification.

UltraBlame original commit: cec8d814aa8326322ce3c47bc5321d6092fb5655