Bug 1384695 - Fetch Android Gradle dependencies from jcenter and maven.google.com. r=sebastian
I took the time to change jcentral (which is just wrong) to jcenter,
which is the tag used in the nexus.xml.

Order matters!  Gradle resolves dependencies in the order given.  That
is, jcenter is preferred to google.

MozReview-Commit-ID: CcWBukhiHa4

UltraBlame original commit: e5ff31e79fc3cd95ce7e45853e250d012103b51d