Bug 1383653 P3 - notify the MediaDecoder::BackgroundVideoDecodingPermissionObserver when a tab is selected; r=dao
In bug 1274919, we only distribute the hovered/un-hovered status to BackgroundVideoDecodingPermissionObserver if a tab is in the background.

Consider the following steps:
Step 1: move mouse to a background tab -> send the hovered event.
Step 2: click on the background tab -> turn it into the foreground.
Step 3: move mouse to leave the tab -> DOES NOT send un-hovered event since the tab is already in the foreground.

In this way, the MediaDecoder is not notified the "un-hovered" event so it keeps wrong states.

In this patch, we reset the state when a tab is selected.

MozReview-Commit-ID: A4kZpwKyQCy

UltraBlame original commit: 8c470a91f0ff7617990e08e09e696612dd674d9e