Bug 1373672 - Part 2: nsDocShell::mChildOffset should be signed, r=smaug
All consumers of this value expect the passed-in value to be signed, and a
negative value is stored into this variable (-1) when the docshell was
dynamically added. It makes more sense for this to be signed.

MozReview-Commit-ID: 8iKDOAx7O2R

UltraBlame original commit: dc318b0913b97dba5f8c8ec2735f96b7d5a30d69