Bug 1385744 - Stop using a cache for the llvm source. r=gps
The build-clang script doesn't handle properly updating the svn tree
when it already exists and the svn url changed (which happens when using
different branches), so if a clang job somehow gets a host that happen
to have run another clang job before, it may not actually build what
was intended to be built.

That event, however, is rather rare, because clang jobs don't happen
regularly enough. Fast iterations on try can cause the problem to
appear, though.

The former makes the use of a cache kind of useless, and the latter
barely benefits from the cache anyways. So let's just solve this the
easy way and remove the use of the cache. The alternative would be to
fiddle with svn switch, and I don't really feel like it.

Eventually, those jobs should use some other source than svn anyways,
one that is more verifiable (tarballs for releases, and the git mirrors
for trunk).

UltraBlame original commit: 22ad2f44b6f841593a6235262bfda0df17373b1d