Bug 1380709 - Add mochitests and mocha tests to ensure entries are loaded as expected. r=bgrins
This adds several tests to make sure we can expand Sets and Maps entries, and that we manage
those properties in the store as expected.
We also add a stub for a Set object and take this as an opportunity to fix the stub name.

MozReview-Commit-ID: ChhIHKMPINF

UltraBlame original commit: 69f4c2e35bb82ed0710fde6783c7a0c8a890ef1c