Bug 1386305 - (DLC) Remove outdated bootstrap catalog. r=mcomella
The bootstrap catalog is outdated and downloading files from it returns in 404 errors from the CDN.
It's impossible to keep the bootstrap catalog updated. Instead let's start with an empty catalog
and rely on the latest catalog that we synchronize from Kinto.

MozReview-Commit-ID: FduZWJBvj16

UltraBlame original commit: 92853fc3f06094dace835f751317ae78c9668439