Bug 1239292 - Remove the gfx job scheduler. r=jrmuizel
We gave up on the iea of parallel tile rendering in favor of different approaches, so the job scheduler isn't used at all. The several seconds spend testing the job scheduler in each push is a waste of infra. In addition, rust is a better avenue to tackle this type of system.

UltraBlame original commit: 009af86a30269949fa6777db130722b975c9605b