Bug 1384120 - Replace old pseudo style context with a new style context including animations. r=emilio
When a reframe happens on the parent of a pseudo element which has animations,
we need to grab style for the pseudo element that includes the animations'
style and also *replace* old style context (that does not include animations'
style) with it. Otherwise, we will use the old style context that has *no*
animations style, as a result, we will see a flicker right after the reframe.

Two reftests in this patch fail without this fix.  One is for CSS transitions,
the other one is for CSS animations.

MozReview-Commit-ID: 6pCdnQ1DGUY

UltraBlame original commit: d79b9e6b0d037e1869563d303430bdf7ce53be6a