Bug 1387482 - default to true when reading pref devtools.inspector.enabled from Firefox;r=mixedpuppy
This preference will be moved inside the devtools addon and won't be always available.
By default consuemrs should consider this to be true, as the inspector is always enabled
by default.

If devtools are not installed, the DevToolsShim will be responsible for redirecting the
call to inspect node to the devtools installation flow.

MozReview-Commit-ID: 279mJ5WRzok

UltraBlame original commit: 1ea77cc3f2e1f308a47b7fa0a8d16167c4869ce4