Bug 1388881 - Use namespace id instead of atom in synthesize_presentational_hints_for_legacy_attributes. r=manishearth
Using namespace id fixes this issue because in Gecko, the pref of MathML
(as well as SVG) works in the way that we choose a different namespace
id (the disabled id) for the elements. Those ids are mapped to the same
namespace atom as normal ids, which means if we use the atom, we would
treat the elements like normal mathml elements.

MozReview-Commit-ID: 9YBBokbP04M

UltraBlame original commit: fdcd393313a80cab8696352f3c9c55666441a9c4