Bug 1380081 - Part 10: Stop running BHR on beta, r=froydnj
These changes are going to increase the amount of data which we collect from BHR
a lot. It would be dangerous to run it on beta, especially considering how soon
the next merge is.

This should turn it off for 100% of beta users if I understand the logic
correctly.

MozReview-Commit-ID: 3HyEKWdXaqU

UltraBlame original commit: d34e59a86079e6f119c38030c4fd00318f118ce4