Bug 1389021 - Release owned CompositorManagerParent in CompositorBridgeParent::DeferredDestroy. r=nical
Prior to bug 1365927 part 5, the compositor thread holder owned by a
CompositorBridgeParent or CrossProcessCompositorBridgeParent object
would be released in its DeferredDestroy method. After part 5 (and
related part 6), the CompositorBridgeParentBase holds a reference to the
CompositorManagerParent which in turn holds a reference to the
compositor thread. We didn't explicitly free the manager in the
DeferredDestroy function which in effect extended the life of the
compositor thread. Now we release the manager in the same place we used
to hold the compositor thread with the aim this eliminates a class of
shutdown hangs where the compositor thread never gets cleaned up.

UltraBlame original commit: dd4bf41ca47d3e3036b194a61af4983c58dd14b2